waffle: use enum list in waffle_enum_to_string()
Use the new WAFFLE_ENUM_LIST macro to implement waffle_enum_to_string()
instead of specifying all the enums again.
Now it will no longer be necessary to update waffle_enum_to_string()
when adding an enum.
Use bsearch() instead of a case statement, not because it's better,
but to be symmetric with a new waffle_string_to_enum() function.

Signed-off-by: Frank Henigman <fjhenigman@google.com>