WIP null: do not create EGLConfig
Add wnull_config to contain the logic for validating a waffle config
for the null platform.
Remove EGLConfig creation - we don't create an EGLSurface so we don't need it.
This doesn't work yet.

On second thought this is a dumb idea.  We still need the config to create
a context, except perhaps on mesa which has configless contexts, but we
want this to work on other implementations than mesa, so forget it.